+++ testing/mocks/user.ts
@@ -6,6 +6,7 @@ import type { LoggedInUser } from 'models';
 export function createMockUser(user?: Partial<LoggedInUser>): LoggedInUser {
   const id = user?.id ?? uuid();
   return {
+    primaryWalletId: null,
     publishToLensDefault: false,
     id: uuid(),
     avatar: null,
